django-shirow/requirements.txt
pyjwt
pypandoc # only for converting README.md
redis
